Migrated to SDL3, replaced ints with floats.
Changes:
Finished migrating to SDL3.

SDL3 replaced a lot of int-based image interfaces with floats, so
it was a good opportunity to do the same with our UI values.

+++ Source/Core/Public/AUI/Core.h
@@ -69,7 +69,7 @@ public:
      *
      * See EventRouter.h member comments for full information on drag and drop.
      */
-    static void setDragTriggerDistance(int newDragTriggerDistance);
+    static void setDragTriggerDistance(float newDragTriggerDistance);
 
     /**
      * If true, a TextInput widget is currently focused and receiving keyboard
@@ -84,7 +84,7 @@ public:
     static ScreenResolution getLogicalScreenSize();
     static ScreenResolution getActualScreenSize();
     static AUI::AssetCache& getAssetCache();
-    static int getSquaredDragTriggerDistance();
+    static float getSquaredDragTriggerDistance();
 
 private:
     /** Friend Widget so it can update the widget count. */
@@ -116,9 +116,9 @@ private:
 
     /** The distance in pixels that the mouse must travel to trigger to a drag
         and drop event. */
-    static int dragTriggerDistance;
+    static float dragTriggerDistance;
     /** The squared dragTriggerDistance, for more efficient calculations. */
-    static int squaredDragTriggerDistance;
+    static float squaredDragTriggerDistance;
 
     /** Keeps a count of the number of currently constructed widgets.
         Used to check if it's safe to Quit(). */
