+++ Source/SharedDefinitions/Public/AUI/Padding.h
@@ -7,10 +7,10 @@ namespace AUI
  * such as the distance between the text and the border of a TextInput.
  */
 struct Padding {
-    int top{0};
-    int right{0};
-    int bottom{0};
-    int left{0};
+    float top{0};
+    float right{0};
+    float bottom{0};
+    float left{0};
 };
 
 } // namespace AUI
